Renamed the engine from Shinobu to Aphelion
Aphelion is the new name for the engine. Shinobu was a placeholder name
and while I like the tone of it, there isn't really anything more to it than that.

Aphelion is the name of the space ship from my favorite childhood game,
Ratchet and Clank. It represents the furthest point to the sun from an orbit.

It was the name that a friend suggested to me. So having him suggest the same
name that I had in mind just confirmed the fact that this should be the new
name of the engine going forward.

So with that out of the way, I present to you, the Aphelion Engine!